src/api/requirements.txt
Flask==1.0.2
Flask-HTTPAuth==3.2.4
pyotp==2.2.7
python-dotenv==0.10.1
